requirements.txt
Django==5.2.10
django-environ==0.12.0
python-decouple==3.8
python-dotenv==1.0.1
dj-database-url==2.3.0
whitenoise==6.8.2
gunicorn==20.1.0
django-allauth==65.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-spectacular==0.27.2
django-debug-toolbar==6.2.0
django-cors-headers==4.3.0
django-sslserver==0.22
channels==4.3.1
daphne==4.2.1
channels_redis==4.3.0
redis==6.4.0
Pillow==10.1.0
cloudinary==1.44.1
django-cloudinary-storage==0.3.0
psycopg2-binary==2.9.10
django-phonenumber-field==8.1.0
phonenumbers==9.0.3
